#362 refactor: AssetType을 CategoryType으로 변환
