+++ src/main/java/com/floney/floney/analyze/dto/process/BookAnalyzer.java
@@ -1,14 +1,18 @@
 package com.floney.floney.analyze.dto.process;
 
 import com.floney.floney.analyze.dto.response.AnalyzeResponseByAsset;
-import com.floney.floney.book.domain.constant.AssetType;
 import com.floney.floney.book.dto.process.AssetInfo;
 
 import java.time.LocalDate;
 import java.util.Map;
 
+import static com.floney.floney.book.domain.category.CategoryType.INCOME;
+import static com.floney.floney.book.domain.category.CategoryType.OUTCOME;
+
 public class BookAnalyzer {
+
     private static final int STANDARD = 0;
+
     private final Map<String, Double> totalExpenses;
 
     public BookAnalyzer(Map<String, Double> totalExpenses) {
@@ -16,10 +20,12 @@ public class BookAnalyzer {
     }
 
     public double differenceInAndOutCome() {
-        return totalExpenses.get(AssetType.INCOME.getKind()) - totalExpenses.get(AssetType.OUTCOME.getKind());
+        return totalExpenses.get(INCOME.getMeaning()) - totalExpenses.get(OUTCOME.getMeaning());
     }
 
-    public AnalyzeResponseByAsset analyzeAsset(String requestMonth, double savedAsset, Map<LocalDate, AssetInfo> assetInfo) {
+    public AnalyzeResponseByAsset analyzeAsset(final String requestMonth,
+                                               final double savedAsset,
+                                               final Map<LocalDate, AssetInfo> assetInfo) {
         double difference = differenceInAndOutCome();
         LocalDate requestDate = LocalDate.parse(requestMonth);
         double currentAsset = assetInfo.get(requestDate).getAssetMoney();
